requirements.txt
beautifulsoup4
requests
openpyxl
pdfminer.six
jaconv
